Replace regl- with pixi-based matrix renderer
