+++ package.json
@@ -81,8 +81,7 @@
     "redux": "~4.0.5",
     "redux-batched-actions": "~0.4.1",
     "redux-freeze": "~0.1.7",
-    "regl": "~1.3.13",
-    "skmeans": "^0.11.3",
+    "skmeans": "~0.11.3",
     "style-object-to-css-string": "~1.0.1",
     "umap-js": "~1.3.2",
     "with-raf": "~1.1.1"
@@ -90,7 +89,6 @@
   "peerDependencies": {
     "pub-sub-es": "~1.2.1",
     "pixi.js": "~5.1.6",
-    "regl": "~1.3.13",
     "umap-js": "~1.3.2"
   },
   "devDependencies": {
